LayoutTests/fast/regions/parsing/webkit-region-fragment-parsing-expected.txt

Tests parsing of webkit-region-fragment property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testCSSText("-webkit-region-fragment: auto") is "auto"
PASS testCSSText("-webkit-region-fragment: initial") is "initial"
PASS testCSSText("-webkit-region-fragment: inherit") is "inherit"
PASS testCSSText("-webkit-region-fragment: break") is "break"
PASS testCSSText("-webkit-region-fragment: 0") is ""
PASS testCSSText("-webkit-region-fragment: -1") is ""
PASS testCSSText("-webkit-region-fragment: 12.5") is ""
PASS testCSSText("-webkit-region-fragment: 1px") is ""
PASS testComputedStyle("auto") is "auto"
PASS testComputedStyle("initial") is "auto"
PASS testComputedStyle("inherit") is "auto"
PASS testComputedStyle("break") is "break"
PASS testComputedStyle("0") is "auto"
PASS testComputedStyle("-1") is "auto"
PASS testComputedStyle("12.5") is "auto"
PASS testComputedStyle("1px") is "auto"
PASS successfullyParsed is true

TEST COMPLETE
